Drop hard dependency on `lp` file extension
`readModelFile` had a hard dependency on `lp` file extensions that was causing it to break when working with other (valid) file formats).

As a result, I've droped `readCPLEXFile` and replaced it with `readFileContents`, which validates against a list of accepted file formats and either throws an exception or reads the file contents into a raw string.

In an effort to reduce code duplication, I've also created the `getValidFormats` function, which is the **only** place in the code that the list of valid formats should be hard-coded. All other functions call  `getValidFormats` to obtain the list.