Skip `gurobi` specific tests if it's not installed
gurobi isn't an easy packge to install, but will break all kinds of tests if it's missing.

To fix this, i've added the `skip_if_not_installed` flag from `testthat` so that anyone who is working on other parts of the code won't have to fight with all the error messages when running the test suite.