add logic to interface xInput tools
These functions are "smart" to detect whether the code is running inside an Alteryx macro, or workflow, or in an external R session. The defaults are used only when  the code is NOT being run in an Alteryx workflow. This is required since both in a macro and in an external R session, these configuration values are empty.